implement race page view model
