+++ source/TrackGenius/TrackGenius.UI.csproj
@@ -36,6 +36,7 @@
   <ItemGroup>
     <ProjectReference Include="..\TrackGenius.Communication\TrackGenius.Communication.csproj" />
     <ProjectReference Include="..\TrackGenius.Const\TrackGenius.Const.csproj" />
+    <ProjectReference Include="..\TrackGenius.Core\TrackGenius.Core.csproj" />
     <ProjectReference Include="..\TrackGenius.Model\TrackGenius.Model.csproj" />
     <ProjectReference Include="..\TrackGenius.Protocol\TrackGenius.Protocol.csproj" />
   </ItemGroup>
